requirements.txt
numpy==1.26.4
pandas==2.2.2
yfinance==0.2.38
matplotlib==3.8.4
requests>=2.32.4
PyYAML>=6.0
